+++ src/test/java/com/cucumber/utils/compare/StringRegexCompareTests.java
@@ -5,8 +5,7 @@ import org.junit.Test;
 
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 public class StringRegexCompareTests {
 
@@ -15,7 +14,7 @@ public class StringRegexCompareTests {
         String expected = "result";
         String actual = "result";
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertTrue(symbols.isEmpty());
     }
 
@@ -24,7 +23,7 @@ public class StringRegexCompareTests {
         int expected = 1;
         int actual = 1;
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertTrue(symbols.isEmpty());
     }
 
@@ -33,9 +32,33 @@ public class StringRegexCompareTests {
         String expected = "~[sym1]";
         int actual = 10;
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertEquals("10", symbols.get("sym1"));
         assertEquals(1, symbols.size());
+    }
+
+    @Test
+    public void compareStringWithManyConsecutiveAssignSymbols() {
+        String expected = "~[prop1]~[prop2]~[prop3]~[prop4]";
+        String actual = "Andorra";
+        StringRegexCompare matcher = new StringRegexCompare(expected, actual);
+        Map<String, Object> symbols = matcher.compare();
+        assertEquals("Andorra", symbols.get("prop1"));
+        assertTrue(symbols.get("prop2").toString().isEmpty());
+        assertTrue(symbols.get("prop3").toString().isEmpty());
+        assertTrue(symbols.get("prop4").toString().isEmpty());
+        assertEquals(4, symbols.size());
+    }
+
+    @Test
+    public void compareStringWithManyAssignSymbolsBetweenNewLines() {
+        String expected = "~[prop1],\n This is a ~[prop2]\n ~[prop3]!";
+        String actual = "Hello,\n This is a world of many nations \n And 7 continents...!";
+        StringRegexCompare matcher = new StringRegexCompare(expected, actual);
+        Map<String, Object> symbols = matcher.compare();
+        assertEquals("Hello", symbols.get("prop1"));
+        assertEquals("world of many nations ", symbols.get("prop2"));
+        assertEquals("And 7 continents...", symbols.get("prop3"));
     }
 
     @Test(expected = AssertionError.class)
@@ -51,7 +74,7 @@ public class StringRegexCompareTests {
         Integer expected = 1;
         int actual = 1;
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertTrue(symbols.isEmpty());
     }
 
@@ -60,7 +83,7 @@ public class StringRegexCompareTests {
         boolean expected = true;
         Boolean actual = true;
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertTrue(symbols.isEmpty());
     }
 
@@ -85,7 +108,7 @@ public class StringRegexCompareTests {
         String expected = "va(lue";
         String actual = "va(lue";
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertTrue(symbols.isEmpty());
     }
 
@@ -94,7 +117,7 @@ public class StringRegexCompareTests {
         String expected = "va.*ue";
         String actual = "va(lue";
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertTrue(symbols.isEmpty());
     }
 
@@ -111,7 +134,7 @@ public class StringRegexCompareTests {
         String expected = "The ~[sym1] is running through the ~[sym2]";
         String actual = "The rabbit is running through the forest";
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertEquals("rabbit", symbols.get("sym1"));
         assertEquals("forest", symbols.get("sym2"));
         assertEquals(2, symbols.size());
@@ -122,7 +145,7 @@ public class StringRegexCompareTests {
         String expected = ".* Rabbit ~[sym1] in the .*";
         String actual = "The Rabbit is running in the forest";
         StringRegexCompare matcher = new StringRegexCompare(expected, actual);
-        Map<String, String> symbols = matcher.compare();
+        Map<String, Object> symbols = matcher.compare();
         assertEquals("is running", symbols.get("sym1"));
         assertEquals(1, symbols.size());
     }
@@ -146,4 +169,78 @@ public class StringRegexCompareTests {
             throw e;
         }
     }
+
+    @Test
+    public void compareStringWithAssignSymbolsAgainstStringWithRegexCharacters() {
+        String expected = "This is regex ~[regex]";
+        String actual = "This is regex a|b|c|d";
+        StringRegexCompare matcher = new StringRegexCompare(expected, actual);
+        Map<String, Object> symbols = matcher.compare();
+        assertEquals("a|b|c|d", symbols.get("regex"));
+        assertEquals(1, symbols.size());
+    }
+
+    @Test
+    public void compareStringWithAssignSymbolsAndRegexAgainstStringWithRegexCharacters() {
+        String expected = ".* is regex ~[regex]";
+        String actual = "This is regex a|b|c|d";
+        StringRegexCompare matcher = new StringRegexCompare(expected, actual);
+        Map<String, Object> symbols = matcher.compare();
+        assertEquals("a|b|c|d", symbols.get("regex"));
+        assertEquals(1, symbols.size());
+    }
+
+    @Test
+    public void compareStringWithAssignSymbolsAndSeparateRegexAgainstStringWithRegexCharacters() {
+        String expected = ".* is regex ~[regex] \\Q[0-9]*\\E";
+        String actual = "This is regex a|b|c|d [0-9]*";
+        StringRegexCompare matcher = new StringRegexCompare(expected, actual);
+        Map<String, Object> symbols = matcher.compare();
+        assertEquals("a|b|c|d", symbols.get("regex"));
+        assertEquals(1, symbols.size());
+    }
+
+    @Test
+    public void compareStringWithAssignSymbolsAndSeparateRegexAgainstStringWithRegexCharacters_negaive() {
+        String expected = ".* is regex ~[regex] [0-9]*";
+        String actual = "This is regex a|b|c|d [0-9]*";
+        try {
+            new StringRegexCompare(expected, actual).compare();
+        } catch (AssertionError e) {
+            assertEquals("\nEXPECTED:\n" +
+                    ".* is regex \\Qa|b|c|d\\E [0-9]*\n" +
+                    "BUT GOT:\n" +
+                    "This is regex a|b|c|d [0-9]*\n", e.getMessage());
+            return;
+        }
+        fail("Values should not match ! But they do...");
+    }
+
+    @Test
+    public void compareStringWithAssignSymbolsAndRegexAgainstStringWithRegexCharacters_negative() {
+        String expected = ".* is regex ~[regex]lorem";
+        String actual = "This is regex a|b|c|d";
+        try {
+            new StringRegexCompare(expected, actual).compare();
+        } catch (AssertionError e) {
+            assertEquals("\n" +
+                    "EXPECTED:\n" +
+                    ".* is regex ~[regex]lorem\n" +
+                    "BUT GOT:\n" +
+                    "This is regex a|b|c|d\n", e.getMessage());
+            return;
+        }
+        fail("Values should not match ! But they do...");
+    }
+
+    @Test
+    public void compareStringWithAssignSymbolsAndInvalidRegexAgainstStringWithRegexCharacters() {
+        String expected = "[ is regex ~[regex]";
+        String actual = "[ is regex a|b|c|d";
+        StringRegexCompare matcher = new StringRegexCompare(expected, actual);
+        Map<String, Object> symbols = matcher.compare();
+        assertEquals("a|b|c|d", symbols.get("regex"));
+        assertEquals(1, symbols.size());
+    }
+
 }