+++ src/test/java/com/cucumber/utils/compare/httpresponsewrapper/HttpResponseWrapperTest.java
@@ -13,7 +13,6 @@ import org.apache.http.protocol.BasicHttpContext;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.io.IOException;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;