+++ src/test/java/com/cucumber/utils/context/props/internal/ScenarioPropsGeneratorTests.java
@@ -1,109 +1,153 @@
-package com.cucumber.utils.placeholders;
+package com.cucumber.utils.context.props.internal;
 
-import com.cucumber.utils.engineering.placeholders.ScenarioPropertiesGenerator;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
-public class PropertiesGeneratorTests {
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+
+public class ScenarioPropsGeneratorTests {
+
 
     @Test
-    public void testPlaceholderGeneratorFromSimpleText() {
-        String a = "~[sym1]";
-        String b = "Moon";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("Moon", generator.getProperties().get("sym1"));
-        assertEquals(1, generator.getProperties().size());
+    public void testPropertiesGeneratorFromObject() {
+        String a = "~[var]";
+        List<String> b = new ArrayList<>();
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals(b, props.get("var"));
+        assertEquals(1, props.size());
     }
 
     @Test
-    public void testPlaceholderGeneratorFromTextWithSpecialCharacters() {
+    public void testPropertiesGeneratorFromNull() {
+        String a = "~[var]";
+        ScenarioPropsGenerator generator = new ScenarioPropsGenerator(a, null).match();
+        Map<String, Object> props = generator.getProperties();
+        assertNull(props.get("var"));
+        assertEquals(1, props.size());
+        assertNull(generator.getSubstitutedSource());
+    }
+
+    @Test
+    public void testPropertiesGeneratorFromSimpleText() {
+        String a = "~[sym1]";
+        String b = "Moon";
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals("Moon", props.get("sym1"));
+        assertEquals(1, props.size());
+    }
+
+    @Test
+    public void testPropertiesGeneratorFromTextWithSpecialCharacters() {
         String a = "~[sym1]";
         String b = "{\"test\":\"M^o|%o$n\"";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals(b, generator.getProperties().get("sym1"));
-        assertEquals(1, generator.getProperties().size());
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals(b, props.get("sym1"));
+        assertEquals(1, props.size());
     }
 
     @Test
-    public void testPlaceholderGeneratorFromSimpleText_negative() {
+    public void testPropertiesGeneratorFromSimpleText_negative() {
         String a = "foo ~[sym1] bar";
         String b = "foo some bra";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals(null, generator.getProperties().get("sym1"));
-        assertEquals(0, generator.getProperties().size());
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertNull(props.get("sym1"));
+        assertEquals(0, props.size());
     }
 
     @Test
-    public void testPlaceholderGeneratorDuplicated() {
-        String a = "~[sym1] ~[sym1]";
-        String b = "Moon Sun";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("Sun", generator.getProperties().get("sym1"));
-        assertEquals(1, generator.getProperties().size());
+    public void testPropertiesGeneratorDuplicates() {
+        String a = "~[sym1] ~[sym1] ~[sym2]";
+        String b = "Moon Sun Tue";
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals("Sun", props.get("sym1"));
+        assertEquals("Tue", props.get("sym2"));
+        assertEquals(2, props.size());
     }
 
     @Test
-    public void testPlaceholderGeneratorFromSimpleTextWithRegex() {
+    public void testPropertiesGeneratorWithTailPlaceholder() {
+        String a = "Lorem Ipsum ~[prop1]";
+        String b = "Lorem Ipsum text";
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals("text", props.get("prop1"));
+        assertEquals(1, props.size());
+    }
+
+    @Test
+    public void testPropertiesGeneratorWithNoMatch() {
+        String a = "this is ~[prop1] property ~[prop2]";
+        String b = "this is ~[prop1] property";
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals(0, props.size());
+    }
+
+    @Test
+    public void testPropertiesGeneratorFromSimpleTextWithRegex() {
         String a = ".*M~[sym1]n.*";
         String b = "Moon";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("oo", generator.getProperties().get("sym1"));
-        assertEquals(1, generator.getProperties().size());
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals("oo", props.get("sym1"));
+        assertEquals(1, props.size());
     }
 
     @Test
-    public void testEmptyPlaceholderGeneratorFromSimpleText() {
+    public void testEmptyPropertiesGeneratorFromSimpleText() {
         String a = "~[sym1]Moon";
         String b = "Moon";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("", generator.getProperties().get("sym1"));
-        assertEquals(1, generator.getProperties().size());
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals("", props.get("sym1"));
+        assertEquals(1, props.size());
     }
 
     @Test
-    public void testEmptyPlaceholderGeneratorFromSimpleTextWithRegex() {
+    public void testEmptyPropertiesGeneratorFromSimpleTextWithRegex() {
         String a = ".*~[sym1]n.*";
         String b = "Moon";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("", generator.getProperties().get("sym1"));
-        assertEquals(1, generator.getProperties().size());
+        Map<String, Object> props = new ScenarioPropsGenerator(a, b).match().getProperties();
+        assertEquals("", props.get("sym1"));
+        assertEquals(1, props.size());
     }
 
     @Test
-    public void testPlaceholderGeneratorFromInvalidRegex() {
+    public void testPropertiesGeneratorFromInvalidRegex() {
         String a = "The ~[var1] is ru.*n(ning through the ~[var2]";
         String b = "something here The rab\nbit is ru.*n(ning through the forest";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("rab\nbit", generator.getProperties().get("var1"));
-        assertEquals("forest", generator.getProperties().get("var2"));
+        ScenarioPropsGenerator generator = new ScenarioPropsGenerator(a, b).match();
+        Map<String, Object> props = generator.getProperties();
+        assertEquals("rab\nbit", props.get("var1"));
+        assertEquals("forest", props.get("var2"));
         assertEquals("The rab\nbit is ru.*n(ning through the forest",
-                generator.getParsedTarget());
+                generator.getSubstitutedSource());
     }
 
     @Test
-    public void testPlaceholderGeneratorFromRegex() {
+    public void testPropertiesGeneratorFromRegex() {
         String a = ".* The ~[var1] is ru\\Q.*\\En.*g through ~[var2] .*";
         String b = "something here The rab\nbit is ru.*nning through the forest";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("rab\nbit", generator.getProperties().get("var1"));
-        assertEquals("the", generator.getProperties().get("var2"));
-        assertEquals(generator.getParsedTarget(),
-                ".* The rab\nbit is ru\\Q.*\\En.*g through the .*",
-                generator.getParsedTarget());
+        ScenarioPropsGenerator generator = new ScenarioPropsGenerator(a, b).match();
+        Map<String, Object> props = generator.getProperties();
+        assertEquals("rab\nbit", props.get("var1"));
+        assertEquals("the", props.get("var2"));
+        assertEquals(generator.getSubstitutedSource().toString(),
+                ".* The rab\nbit is ru\\Q.*\\En.*g through the .*", generator.getSubstitutedSource());
     }
 
     @Test
-    public void testPlaceholderGeneratorFromSimpleJson() {
+    public void testPropertiesGeneratorFromSimpleJson() {
         String a = "{\"a\":[1,~[var1],3,4,5],\"b\":{\"k\":\"i\"}}";
         String b = "{\"a\":[1,2,3,4,5],\"b\":{\"k\":\"i\"}}";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("2", generator.getProperties().get("var1"));
-        assertEquals(b, generator.getParsedTarget());
+        ScenarioPropsGenerator generator = new ScenarioPropsGenerator(a, b).match();
+        Map<String, Object> props = generator.getProperties();
+        assertEquals("2", props.get("var1"));
+        assertEquals(b, generator.getSubstitutedSource());
     }
 
     @Test
-    public void testPlaceholderGeneratorFromJson() {
+    public void testPropertiesGeneratorFromJson() {
         String a = "[\n" + "  {\n" + "    \"_id\": \"5b48fcdd5c3da373f5114ec7\",\n"
                 + "    \"index\": 0,\n"
                 + "    \"guid\": \"ed43e5b1-3da2-4ebe-baa5-e44d4eb18a5c\",\n"
@@ -150,11 +194,12 @@ public class PropertiesGeneratorTests {
                 + "        \"name\": \"Dena \nSosa\"\n" + "      }\n" + "    ],\n"
                 + "    \"greeting\": \"Hello, Tonya Schneider! You have 9 unread messages.\",\n"
                 + "    \"favoriteFruit\": \"banana\"\n" + "  }\n" + "]";
-        ScenarioPropertiesGenerator generator = new ScenarioPropertiesGenerator(a, b);
-        assertEquals("Levine", generator.getProperties().get("var1"));
-        assertEquals("favoriteFruit", generator.getProperties().get("var2"));
-        assertEquals("consect(etur", generator.getProperties().get("var3"));
-        assertEquals("na \nSo", generator.getProperties().get("var4"));
-        assertEquals(b, generator.getParsedTarget());
+        ScenarioPropsGenerator generator = new ScenarioPropsGenerator(a, b).match();
+        Map<String, Object> props = generator.getProperties();
+        assertEquals("Levine", props.get("var1"));
+        assertEquals("favoriteFruit", props.get("var2"));
+        assertEquals("consect(etur", props.get("var3"));
+        assertEquals("na \nSo", props.get("var4"));
+        assertEquals(b, generator.getSubstitutedSource());
     }
 }